[UDT-86] refactor: 목데이터 및 UI 수정
